Isolate path resolution prepare URI for exec
