+++ .obj/parse/header.d
@@ -1,0 +1,34 @@
+.obj/parse/header.o: src/parse/header.cpp src/parse/../../inc/webserv.hpp \
+  inc/webserv_enum.hpp inc/webserv_utils.hpp inc/parse/parse.hpp \
+  inc/parse/parse_utils.hpp inc/parse/Parser.hpp inc/IModule.hpp \
+  inc/webserv.hpp inc/read/Reader.hpp inc/read/../webserv.hpp \
+  inc/validate/validate.hpp inc/validate/Validator.hpp \
+  inc/validate/../webserv.hpp inc/serverConfig.hpp
+
+src/parse/../../inc/webserv.hpp:
+
+inc/webserv_enum.hpp:
+
+inc/webserv_utils.hpp:
+
+inc/parse/parse.hpp:
+
+inc/parse/parse_utils.hpp:
+
+inc/parse/Parser.hpp:
+
+inc/IModule.hpp:
+
+inc/webserv.hpp:
+
+inc/read/Reader.hpp:
+
+inc/read/../webserv.hpp:
+
+inc/validate/validate.hpp:
+
+inc/validate/Validator.hpp:
+
+inc/validate/../webserv.hpp:
+
+inc/serverConfig.hpp:
