+++ inc/validate/Validator.hpp
@@ -1,5 +1,6 @@
 #pragma once
 #include "../webserv.hpp"
+#include <vector>
 
 class Validator {
 private:
@@ -11,8 +12,11 @@ private:
 	//Worst case we skip the body
 	void validateBenign(Conversation& conv);
 	void validateUri(Conversation& conv);
-	void parsePath(Conversation& conv);
-	size_t matchLoc(Conversation& conv);
+	void segmentLinter(Conversation& conv, std::vector<std::string>& seg);
+	void queryLinter(Conversation& conv);
+	std::vector<std::string> parsePath(Conversation& conv);
+	size_t matchLoc(Conversation& conv, std::vector<std::string>& seg);
+	void assembleUri(Conversation& conv, std::vector<std::string>& seg, size_t match);
 	void stripHost(Conversation& conv);
 	void validateMethod(Conversation& conv);
