+++ inc/validate/validate.hpp
@@ -5,9 +5,4 @@
 const std::string pchar = alpha + base10 + "-._~!$&'()*+,;=";
 const std::string qchar = pchar + "?/";
 
-struct parsedPath {
-	std::vector<std::string> seg;
-	bool hasQuery;
-	std::string query;
-};
 void skipBody(Conversation& conv, statusCode status);
